// src/types/IAsset.d.ts
interface IAsset {
	ID: number;
	TYPE: string;
	ID_LEGACY: 1182;
	SYMBOL: string;
	URI: string;
	ASSET_TYPE: string;
	CREATED_ON: number;
	UPDATED_ON: number;
	PUBLIC_NOTICE: null;
	NAME: string;
	LOGO_URL: string;
	LAUNCH_DATE: number;
	PREVIOUS_ASSET_SYMBOLS: null;
	PRICE_CHANGE_FLAG?: 'UP' | 'DOWN';
	ASSET_ALTERNATIVE_IDS: {
		NAME: string;
		ID: string;
	}[];
	ASSET_DESCRIPTION_SNIPPET: string;
	ASSET_DECIMAL_POINTS: number;
	SUPPLY_MAX: number;
	SUPPLY_ISSUED: number;
	SUPPLY_TOTAL: number;
	SUPPLY_CIRCULATING: number;
	SUPPLY_FUTURE: number;
	SUPPLY_LOCKED: number;
	SUPPLY_BURNT: number;
	SUPPLY_STAKED: number;
	LAST_BLOCK_MINT: number;
	LAST_BLOCK_BURN: null;
	PRICE_USD: number;
	PRICE_USD_SOURCE: string;
	PRICE_USD_LAST_UPDATE_TS: number;
	PRICE_CONVERSION_ASSET: {
		ID: number;
		SYMBOL: string;
		ASSET_TYPE: string;
	};
	PRICE_CONVERSION_RATE: number;
	PRICE_CONVERSION_VALUE: number;
	PRICE_CONVERSION_SOURCE: string;
	PRICE_CONVERSION_LAST_UPDATE_TS: number;
	MKT_CAP_PENALTY: number;
	CIRCULATING_MKT_CAP_USD: number;
	TOTAL_MKT_CAP_USD: number;
	CIRCULATING_MKT_CAP_CONVERSION: number;
	TOTAL_MKT_CAP_CONVERSION: number;
	SPOT_MOVING_24_HOUR_QUOTE_VOLUME_TOP_TIER_DIRECT_USD: number;
	SPOT_MOVING_24_HOUR_QUOTE_VOLUME_DIRECT_USD: number;
	SPOT_MOVING_24_HOUR_QUOTE_VOLUME_TOP_TIER_USD: number;
	SPOT_MOVING_24_HOUR_QUOTE_VOLUME_USD: number;
	SPOT_MOVING_24_HOUR_QUOTE_VOLUME_TOP_TIER_CONVERSION: number;
	SPOT_MOVING_24_HOUR_QUOTE_VOLUME_CONVERSION: number;
	SPOT_MOVING_7_DAY_QUOTE_VOLUME_TOP_TIER_DIRECT_USD: number;
	SPOT_MOVING_7_DAY_QUOTE_VOLUME_DIRECT_USD: number;
	SPOT_MOVING_7_DAY_QUOTE_VOLUME_TOP_TIER_USD: number;
	SPOT_MOVING_7_DAY_QUOTE_VOLUME_USD: number;
	SPOT_MOVING_7_DAY_QUOTE_VOLUME_TOP_TIER_CONVERSION: number;
	SPOT_MOVING_7_DAY_QUOTE_VOLUME_CONVERSION: number;
	SPOT_MOVING_30_DAY_QUOTE_VOLUME_TOP_TIER_DIRECT_USD: number;
	SPOT_MOVING_30_DAY_QUOTE_VOLUME_DIRECT_USD: number;
	SPOT_MOVING_30_DAY_QUOTE_VOLUME_TOP_TIER_USD: number;
	SPOT_MOVING_30_DAY_QUOTE_VOLUME_USD: number;
	SPOT_MOVING_30_DAY_QUOTE_VOLUME_TOP_TIER_CONVERSION: number;
	SPOT_MOVING_30_DAY_QUOTE_VOLUME_CONVERSION: number;
	SPOT_MOVING_24_HOUR_CHANGE_USD: number;
	SPOT_MOVING_24_HOUR_CHANGE_PERCENTAGE_USD: number;
	SPOT_MOVING_24_HOUR_CHANGE_CONVERSION: number;
	SPOT_MOVING_24_HOUR_CHANGE_PERCENTAGE_CONVERSION: number;
	SPOT_MOVING_7_DAY_CHANGE_USD: number;
	SPOT_MOVING_7_DAY_CHANGE_PERCENTAGE_USD: number;
	SPOT_MOVING_7_DAY_CHANGE_CONVERSION: number;
	SPOT_MOVING_7_DAY_CHANGE_PERCENTAGE_CONVERSION: number;
	SPOT_MOVING_30_DAY_CHANGE_USD: number;
	SPOT_MOVING_30_DAY_CHANGE_PERCENTAGE_USD: number;
	SPOT_MOVING_30_DAY_CHANGE_CONVERSION: number;
	SPOT_MOVING_30_DAY_CHANGE_PERCENTAGE_CONVERSION: number;
	TOPLIST_BASE_RANK: {
		CREATED_ON: number;
		LAUNCH_DATE: number;
		CIRCULATING_MKT_CAP_USD: number;
		TOTAL_MKT_CAP_USD: number;
		SPOT_MOVING_24_HOUR_QUOTE_VOLUME_USD: number;
		SPOT_MOVING_7_DAY_QUOTE_VOLUME_USD: number;
		SPOT_MOVING_30_DAY_QUOTE_VOLUME_USD: number;
	};
}

export default IAsset;